SP refactor design

+++ src/app/card-element-details/card-element-details.component.ts
@@ -3,6 +3,8 @@ import { CardElement } from '../card-element';
 import { EventService } from '../event.service';
 import { CardElementService } from '../card-element.service';
 import { Router } from "@angular/router";
+import { ModalParams } from '../modal/modalClass';
+import { ModalService } from '../modal/modal.service';
 
 @Component({
     selector: 'app-card-element-details',
@@ -12,8 +14,12 @@ import { Router } from "@angular/router";
 export class CardElementDetailsComponent implements OnInit {
     public cardElement: CardElement;
     public votersValues: any[];
+    public positiveVoters : any[];
+    public negativeVoters : any[];
+    public maxTableSize: number;
+    public tableSizeRangeArray: number [];
 
-    constructor(private eventService: EventService, private cardElementService: CardElementService, private router: Router) {
+    constructor(private eventService: EventService, private cardElementService: CardElementService, private router: Router, private modalService: ModalService) {
         this.eventService.cardElementDeleted$.subscribe(value => {
             if (value) {
                 this.navigateToEvent();
@@ -25,6 +31,14 @@ export class CardElementDetailsComponent implements OnInit {
         this.cardElement = this.eventService.cardElementToShowDetails;
         this.votersValues = [];
         this.sortCardElementVoter();
+        this.handlePositiveNegativeVotes();
+    }
+
+    public handlePositiveNegativeVotes() {
+        this.positiveVoters = this.cardElement.voters.filter( x => x.nbVotes == 1);
+        this.negativeVoters = this.cardElement.voters.filter( x => x.nbVotes == 2);
+        this.maxTableSize = Math.max(this.positiveVoters.length, this.negativeVoters.length);
+        this.tableSizeRangeArray = Array.from({length: this.maxTableSize}, (x, i) => i);
     }
 
     public sortCardElementVoter(): void {
@@ -47,7 +61,13 @@ export class CardElementDetailsComponent implements OnInit {
     }
 
     public deleteCardElement() {
-        this.eventService.deleteCardElement(this.cardElement);
+        let modalData = new ModalParams();
+        modalData.setQuestion("Voulez vous vraiment supprimer cette carte ?", 'Attention, cela est irréversible', 'Oui');
+        this.modalService.listenToNewValueAndOpenModal(this, modalData, (valueReturned) => {
+            if (valueReturned) {
+                this.eventService.deleteCardElement(this.cardElement);
+            }
+        });
     }
 
     public navigateToEvent(): void {
